@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.20)
 project(Tetris)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 add_subdirectory(dependencies/fmt)
 
 find_package(SFML COMPONENTS system window graphics network audio REQUIRED)
